chore(typescript_indexer): extract plugin API to a separate file
